Refactor + introduce Anthropic support.
