+++ notdiamond/_init.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-import gc
-from typing import Dict, List, Union
+from typing import Any, Dict, List, Union
 
 from openai import AsyncOpenAI, OpenAI
 
@@ -10,6 +9,8 @@ from notdiamond.toolkit.openai import OpenAIRetryWrapper
 
 
 def init(
+    # todo [a9]: update annotation for correct client types
+    client: OpenAI | AsyncOpenAI | Any,
     models: Union[Dict[str | LLMConfig, float], List[str | LLMConfig]],
     # todo [a9]: accept dict of model -> max_retries
     max_retries: int,
@@ -25,6 +26,7 @@ def init(
     ```python
         openai_client = OpenAI(...)
         notdiamond.init(
+            openai_client,
             models={"gpt-3.5-turbo": 0.9, "claude-3-5-sonnet-20240620": 0.1},
             max_retries=3,
             timeout=10.0,
@@ -37,16 +39,16 @@ def init(
         )
     ```
     """
-    openai_client = gc.get_objects(type_filter=OpenAI)
-    if not openai_client:
-        openai_client = gc.get_objects(type_filter=AsyncOpenAI)
+    # openai_client = gc.get_objects(type_filter=OpenAI)
+    # if not openai_client:
+    #     openai_client = gc.get_objects(type_filter=AsyncOpenAI)
 
-    if not openai_client:
-        raise ValueError(
-            "No OpenAI or AsyncOpenAI client found. Is this correct?"
-        )
+    # if not openai_client:
+    #     raise ValueError(
+    #         "No OpenAI or AsyncOpenAI client found. Is this correct?"
+    #     )
 
     client_wrapper = OpenAIRetryWrapper(
-        openai_client, models, max_retries, timeout, fallback
+        client, models, max_retries, timeout, fallback
     )
     return client_wrapper
